requirements.txt
Flask>=2.3.2
flask-login>=0.6.2
flask-wtf>=1.2.1
cryptography>=42.0.0
pyotp>=2.9.0
langchain>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
